Not complete, but progress
Added some outstanding items to both the watch and the action items. Struggling to get the observation portion working. every time i try to add that in, it magically stops doing the discharge part. So I am working that. This is a working version without the observations, if you wanted to review/play.